+++ react-project/src/components/ErrorMessage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./ErrorMessage.css";
+import "./css/ErrorMessage.css";
 
 interface iErrorM {
   message: string;
