+++ react-project/src/components/Loader.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import "./spinner.css";
+import "./css/spinner.css";
 
 export const Loader = () => {
   return (
-    <div className="spinner-container ">
+    <div className="spinner-container">
       <div className="loading-spinner"></div>
     </div>
   );
