+++ react-project/src/components/Modal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Modal.css";
+import "./css/Modal.css";
 
 interface ModalProps {
   children: React.ReactNode;
