+++ react-project/src/pages/About.tsx
@@ -1,5 +1,5 @@
-import "./About.css";
 import React from "react";
+import "./css/About.css";
 
 export function AboutUs() {
   return (
