+++ react-project/src/pages/Form.tsx
@@ -1,29 +1,16 @@
-import "./Form.css";
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
+import { useForm, SubmitHandler, Controller } from "react-hook-form";
 import { imgValid, lengthValid } from "../utils/checkFormValid";
 import { ErrorMessage } from "../components/ErrorMessage";
 import { FormCards } from "../components/FormCards";
-import { useForm, SubmitHandler, Controller } from "react-hook-form";
 import { ICardValues, IFormValues } from "../interfaces/formInterfaces";
 import { composeNewCard } from "../utils/composeNewCard";
 import { IRootState } from "../interfaces/store";
 import { changeFormCard } from "../store/sliceFormCard";
+import "./css/Form.css";
 
 export const Form = () => {
-  // const search = useSelector((state: IRootState) => state.search.search);
-  // const [state, setState] = useState(search);
-  // const dispatch = useDispatch();
-
-  // const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-  //   setState(event.target.value);
-  // };
-
-  // const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-  //   if (event.key === "Enter") {
-  //     dispatch(changeSearch({ search: state }));
-  //   }
-  // };
   const dispatch = useDispatch();
   const cardList = useSelector((state: IRootState) => state.formCard.formCard);
   const {
@@ -40,7 +27,6 @@ export const Form = () => {
     const newData = { ...data, image: URL.createObjectURL(data.image![0]) };
 
     const newCard: ICardValues = composeNewCard(newData);
-    // setCardList([...cardList, newCard]);
     dispatch(changeFormCard({ formCard: [...cardList, newCard] }));
     alert("Card was added");
     reset();
