+++ react-project/src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Main.css";
+import "./css/Main.css";
 import { SearchBar } from "../components/SearchBar";
 import { Cards } from "../components/Cards";
