+++ react-project/src/utils/fetchData.ts
@@ -25,18 +25,3 @@ export const LORApi = createApi({
 });
 
 export const { useGetCharactersQuery } = LORApi;
-
-// export const fetchData = async (name: string) => {
-//   let params = `?name=/${name}/i&limit=9`;
-//   if (!name) {
-//     params = `?limit=9`;
-//   }
-//   const rawPersons = await fetch(
-//     `https://the-one-api.dev/v2/character${params}`,
-//     {
-//       headers: headers,
-//     }
-//   );
-//   const persons = await rawPersons.json();
-//   return persons;
-// };
